pyproject.toml
[project]
name = "vlm-safety"
version = "0.1.0"
description = "VLM safety research: adversarial image generation via CLIP embedding optimization"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "langchain-openai>=1.1.10",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "open-clip-torch>=3.2.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vlm_safety"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
